requirements.txt
mlflow
python-dotenv
pandas
numpy
scikit-learn
torch
pydantic
kagglehub